+++ integration_tests/juniper_tests/src/lib.rs
@@ -15,4 +15,16 @@ mod issue_371;
 #[cfg(test)]
 mod issue_398;
 #[cfg(test)]
+mod issue_407;
+#[cfg(test)]
 mod issue_500;
+#[cfg(test)]
+mod issue_914;
+#[cfg(test)]
+mod issue_922;
+#[cfg(test)]
+mod issue_925;
+#[cfg(test)]
+mod issue_945;
+#[cfg(test)]
+mod pre_parse;