+++ juniper/src/executor/look_ahead.rs
@@ -96,6 +96,7 @@ where
     }
 }
 
+#[doc(hidden)]
 #[derive(Debug, Clone, PartialEq)]
 pub struct ChildSelection<'a, S: 'a> {
     pub(super) inner: LookAheadSelection<'a, S>,
@@ -340,14 +341,11 @@ pub trait LookAheadMethods<'sel, S> {
 
     /// Get the the child selection for a given field
     /// If a child has an alias, it will only match if the alias matches `name`
-    #[deprecated(note = "please use `children` to access the child selections instead")]
     fn select_child(&self, name: &str) -> Option<&Self>;
 
     /// Check if a given child selection with a name exists
     /// If a child has an alias, it will only match if the alias matches `name`
-    #[deprecated(note = "please use `children` to access the child selections instead")]
     fn has_child(&self, name: &str) -> bool {
-        #[allow(deprecated)]
         self.select_child(name).is_some()
     }
 
@@ -366,7 +364,6 @@ pub trait LookAheadMethods<'sel, S> {
     }
 
     /// Get the (possibly aliased) names of the top level children for the current selection
-    #[deprecated(note = "please use `children` to access the child selections instead")]
     fn child_names(&self) -> Vec<&'sel str>;
 
     /// Get an iterator over the children for the current selection
@@ -443,7 +440,7 @@ impl<'a, S> LookAheadMethods<'a, S> for LookAheadSelection<'a, S> {
 mod tests {
     use super::*;
     use crate::{
-        ast::Document,
+        ast::{Document, OwnedDocument},
         parser::UnlocatedParseResult,
         schema::model::SchemaType,
         validation::test_harness::{MutationRoot, QueryRoot, SubscriptionRoot},
@@ -451,7 +448,7 @@ mod tests {
     };
     use std::collections::HashMap;
 
-    fn parse_document_source<S>(q: &str) -> UnlocatedParseResult<Document<S>>
+    fn parse_document_source<S>(q: &str) -> UnlocatedParseResult<OwnedDocument<S>>
     where
         S: ScalarValue,
     {
@@ -1307,7 +1304,6 @@ query Hero {
     }
 
     #[test]
-    #[allow(deprecated)]
     fn check_select_child() {
         let lookahead: LookAheadSelection<DefaultScalarValue> = LookAheadSelection {
             name: "hero",
@@ -1459,7 +1455,6 @@ fragment heroFriendNames on Hero {
     }
 
     #[test]
-    #[allow(deprecated)]
     fn check_visitability() {
         let docs = parse_document_source::<DefaultScalarValue>(
             "