+++ juniper/src/types/pointers.rs
@@ -94,10 +94,7 @@ where
     T: FromInputValue<S>,
 {
     fn from_input_value(v: &InputValue<S>) -> Option<Box<T>> {
-        match <T as FromInputValue<S>>::from_input_value(v) {
-            Some(v) => Some(Box::new(v)),
-            None => None,
-        }
+        <T as FromInputValue<S>>::from_input_value(v).map(Box::new)
     }
 }
 
@@ -289,10 +286,7 @@ where
     T: FromInputValue<S>,
 {
     fn from_input_value(v: &InputValue<S>) -> Option<Arc<T>> {
-        match <T as FromInputValue<S>>::from_input_value(v) {
-            Some(v) => Some(Arc::new(v)),
-            None => None,
-        }
+        <T as FromInputValue<S>>::from_input_value(v).map(Arc::new)
     }
 }