+++ juniper/src/validation/multi_visitor.rs
@@ -11,6 +11,7 @@ use crate::{
 #[doc(hidden)]
 pub struct MultiVisitorNil;
 
+#[doc(hidden)]
 impl MultiVisitorNil {
     pub fn with<V>(self, visitor: V) -> MultiVisitorCons<V, Self> {
         MultiVisitorCons(visitor, self)
@@ -229,19 +230,11 @@ where
         self.1.exit_list_value(ctx, l);
     }
 
-    fn enter_object_value(
-        &mut self,
-        ctx: &mut ValidatorContext<'a, S>,
-        o: Spanning<&'a Vec<(Spanning<String>, Spanning<InputValue<S>>)>>,
-    ) {
+    fn enter_object_value(&mut self, ctx: &mut ValidatorContext<'a, S>, o: SpannedObject<'a, S>) {
         self.0.enter_object_value(ctx, o);
         self.1.enter_object_value(ctx, o);
     }
-    fn exit_object_value(
-        &mut self,
-        ctx: &mut ValidatorContext<'a, S>,
-        o: Spanning<&'a Vec<(Spanning<String>, Spanning<InputValue<S>>)>>,
-    ) {
+    fn exit_object_value(&mut self, ctx: &mut ValidatorContext<'a, S>, o: SpannedObject<'a, S>) {
         self.0.exit_object_value(ctx, o);
         self.1.exit_object_value(ctx, o);
     }
@@ -263,3 +256,5 @@ where
         self.1.exit_object_field(ctx, f);
     }
 }
+
+type SpannedObject<'a, S> = Spanning<&'a Vec<(Spanning<String>, Spanning<InputValue<S>>)>>;