+++ juniper_rocket_async/examples/rocket_server.rs
@@ -2,7 +2,7 @@ use juniper::{
     tests::fixtures::starwars::schema::{Database, Query},
     EmptyMutation, EmptySubscription, RootNode,
 };
-use rocket::{response::content, State};
+use rocket::{response::content, Rocket, State};
 
 type Schema = RootNode<'static, Query, EmptyMutation<Database>, EmptySubscription<Database>>;
 
@@ -13,25 +13,25 @@ fn graphiql() -> content::Html<String> {
 
 #[rocket::get("/graphql?<request>")]
 fn get_graphql_handler(
-    context: State<Database>,
+    context: &State<Database>,
     request: juniper_rocket_async::GraphQLRequest,
-    schema: State<Schema>,
+    schema: &State<Schema>,
 ) -> juniper_rocket_async::GraphQLResponse {
-    request.execute_sync(&schema, &context)
+    request.execute_sync(&*schema, &*context)
 }
 
 #[rocket::post("/graphql", data = "<request>")]
 fn post_graphql_handler(
-    context: State<Database>,
+    context: &State<Database>,
     request: juniper_rocket_async::GraphQLRequest,
-    schema: State<Schema>,
+    schema: &State<Schema>,
 ) -> juniper_rocket_async::GraphQLResponse {
-    request.execute_sync(&schema, &context)
+    request.execute_sync(&*schema, &*context)
 }
 
 #[rocket::main]
 async fn main() {
-    rocket::ignite()
+    Rocket::build()
         .manage(Database::new())
         .manage(Schema::new(
             Query,