+++ juniper_warp/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 - Compatibility with the latest `juniper`.
 
+# [[0.6.4] 2021-06-07](https://github.com/graphql-rust/juniper/releases/tag/juniper_warp-0.6.4)
+
+- Compatibility with the latest `juniper`.
+
+# [[0.6.3] 2021-04-03](https://github.com/graphql-rust/juniper/releases/tag/juniper_warp-0.6.3)
+
+- Compatibility with the latest `juniper`.
+
 # [[0.6.2] 2021-01-27](https://github.com/graphql-rust/juniper/releases/tag/juniper_warp-0.6.2)
 
 - Compatibility with the latest `juniper`.